Split off astro style classroom table. Working auto-create for assignments
